requirements.txt
Django==1.4
dj-database-url==0.2.1
django-geoposition==0.1.4
docutils==0.10
gunicorn==0.17.4
lxml==3.1.1
mimeparse==0.1.3
pykml==0.1.0
python-dateutil==1.5
virtualenv==1.9.1
django-tastypie==0.9.12